services/auth/requirements.txt
pydantic==1.10.2
fastapi==0.92.0
uvicorn==0.20.0
orjson==3.8.7
faker==17.6.0
pyjwt==2.6.0
requests==2.28.2
aiohttp==3.8.1
